Fix: Repository 계층 메서드에 @Param 추가, COUNT, SUM 반환 형식에 따른 타입 Long으로 변경(#106)

+++ src/main/java/com/dnd/Exercise/domain/userField/dto/response/TopPlayerDto.java
@@ -16,9 +16,9 @@ public class TopPlayerDto {
         this.value = 0;
     }
 
-    public void updateIfLarger(int newValue, String newName) {
+    public void updateIfLarger(Long newValue, String newName) {
         if (this.value < newValue) {
-            this.value = newValue;
+            this.value = Math.toIntExact(newValue);
             this.name = newName;
         }
     }
